requirements.txt
pytest
pytest-mock
pydantic >= 2.10.6
srt
tqdm >= 4.66.3
ffmpeg-python
inflect